### 2022-07-18 #4586 @editorialbot
[PRE REVIEW]: Pyripherals: A Python Package for Communicating with Peripheral Electronic Devices
**Submitting author:** @Ajstros (<a href="http://orcid.org/0000-0001-9412-2049">Abraham Stroschein</a>)
**Repository:** https://github.com/Ajstros/pyripherals
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.2
