### 2022-07-26 7015f8f
updated SPI clock rate in impedance analyzer example; changed DDR data_to_names for moving ADS['B'] channel to be synced with 'A'; fix bug in ADS8686 lpf Register position
