### 2022-07-26 #4586 @Kevin-Mattheus-Moerman
@jarvist would this work be too far off topic for you to handle as editor?
