### 2022-07-26 #4586 @jarvist
A bit too far away I'm afraid; sounds like more of a job for an electrical engineer, or at least an expt physicist who does a lot of hardware interaction.
