### 2022-07-26 2a7945d
add flag to data_to_names to allow for analysis of old captures
