+++ paper/paper.bib
@@ -1,18 +1,23 @@
 
-@misc{DesignresourcesHDLComponentsI2CController,
-	title = {I2CController},
-	url = {https://github.com/opalkelly-opensource/design-resources/tree/main/HDLComponents/I2CController},
-	urldate = {2022-07-05},
-	year = {2021},
-	publisher = {Opal Kelly}
-}
+@article{leibrandtOpenSourceDigital2015,
+	title = {An open source digital servo for {AMO} physics experiments},
+	url = {http://arxiv.org/abs/1508.06319},
+	abstract = {We describe a general purpose digital servo optimized for feedback control of lasers in atomic, molecular, and optical (AMO) physics experiments. The servo is capable of feedback bandwidths up to roughly 1{\textasciitilde}MHz (limited by the 320{\textasciitilde}ns total latency); loop filter shapes up to fifth order; multiple-input, multiple-output control; and automatic lock acquisition. The configuration of the servo is controlled via a graphical user interface, which also provides a rudimentary software oscilloscope and tools for measurement of system transfer functions. We illustrate the functionality of the digital servo by describing its use in two example scenarios: frequency control of the laser used to probe the narrow clock transition of \${\textasciicircum}\{27\}\$Al\${\textasciicircum}+\$ in an optical atomic clock, and length control of a cavity used for resonant frequency doubling of a laser.},
+	language = {en},
+	urldate = {2021-01-21},
+	journal = {arXiv:1508.06319 [physics]},
+	author = {Leibrandt, David R. and Heidecker, Jason},
+	month = nov,
+	year = {2015},
+	note = {arXiv: 1508.06319},
+	keywords = {Physics - Instrumentation and Detectors, Physics - Atomic Physics}}
 
 @misc{srotOverviewSPIController,
-	title = {{SPI} controller core},
+	title = {Overview :: {SPI} controller core :: {OpenCores}},
 	url = {https://opencores.org/projects/spi},
 	urldate = {2022-07-05},
-	year = {2016},
 	author = {Srot, Simon},
+	year={2016}
 }
 
 @misc{opalkelly.comFrontPanel,
@@ -26,19 +31,100 @@
 	publisher = {Opal Kelly},
 }
 
- @misc{dac80508,
- 	title={{DAC}80508},
-	url={https://www.ti.com/product/DAC80508},
-	journal={DAC80508 data sheet, product information and support | TI.com},
+@article{yuPerformanceLimitationsFPGAbased2018,
+	title = {The performance and limitations of {FPGA}-based digital servos for atomic, molecular, and optical physics experiments},
+	volume = {89},
+	issn = {0034-6748},
+	url = {https://aip.scitation.org/doi/abs/10.1063/1.5001312},
+	doi = {10.1063/1.5001312},
+	number = {2},
+	urldate = {2022-07-28},
+	journal = {Review of Scientific Instruments},
+	author = {Yu, Shi Jing and Fajeau, Emma and Liu, Lin Qiao and Jones, David J. and Madison, Kirk W.},
+	month = feb,
 	year = {2018},
-	urldate = {2022-07-05},
-	publisher={Texas Instruments}
+	note = {Publisher: American Institute of Physics},
+	pages = {025107},
+	file = {Submitted Version:/Users/koer2434/Zotero/storage/D7H4CEJU/Yu et al. - 2018 - The performance and limitations of FPGA-based digi.pdf:application/pdf},
 }
 
[…]
