### 2022-08-02 dbaaad7
Replace twos_comp with faster binary_twos_comp everywhere
