### 2022-08-02 8b36635
Replace missing twos_comp swap
