+++ python/src/pyripherals/utils.py
@@ -193,7 +193,7 @@ def to_voltage(data, num_bits, voltage_range, use_twos_comp=False):
 
     bit_voltage = voltage_range / (2 ** num_bits)
     if use_twos_comp:
-        twos_data = binary_twos_comp(data=data, num_bits=num_bits)
+        twos_data = twos_comp(data=data, num_bits=num_bits)
         data = np.where(np.array(data) >= (1 << num_bits - 1), -1 * twos_data, twos_data)
 
     if type(data) is np.ndarray:
@@ -247,7 +247,7 @@ def from_voltage(voltage, num_bits, voltage_range, with_negatives=False):
         raise TypeError(f'from_voltage voltage expected np.integer, np.floating, list, or np.ndarray type, got {type(voltage)}')
 
     if with_negatives:
-        data = binary_twos_comp(data=data, num_bits=num_bits)
+        data = twos_comp(data=data, num_bits=num_bits)
 
     # Since this system may overflow to 0 on the maximum value, we limit any
     # input voltages of maximum to full-scale.
