### 2022-08-03 314d310
Replace twos_comp everywhere
