### 2022-08-03 d0b4ec6
Fix int_to_custom_signed
