### 2022-08-03 77327af
Incorporate FIFO reset and PIPE read into i2c_receive pipe option
