### 2022-08-03 ed4081e
Add error message for missing endpoints in i2c_receive
