### 2022-08-03 98f40d5
Update docstrings, comments
