### 2022-08-03 3e5b051
Replace twos_comp (#16)
* Switch data_to_names to use faster binary_twos_comp

* Replace twos_comp with faster binary_twos_comp everywhere

* Replace missing twos_comp swap

* Replace twos_comp with custom_signed_to_int and int_to_custom_signed in utils

* Replace twos_comp everywhere

* Fix int_to_custom_signed

* Fix to_voltage bug

* Update docstrings, comments
