### 2022-08-03 9753135
Updated i2c_receive and i2c_read_long docstrings
