### 2022-08-17 c174551
Add bitfile_version to FPGA class
