### 2022-08-19 66795f6
Ads sync (#19) fixes #14
* updated SPI clock rate in impedance analyzer example; changed DDR data_to_names for moving ADS['B'] channel to be synced with 'A'; fix bug in ADS8686 lpf Register position

* add flag to data_to_names to allow for analysis of old captures

* fix missing parameter for data_to_names

Co-authored-by: Abe Stroschein <69999361+Ajstros@users.noreply.github.com>
