### 2022-08-19 fd36347
Add str_bitfile_version
