### 2022-08-19 48bc863
Account for KeyError when accessing BITFILE_VERSION Endpoint
