### 2022-08-19 56e3aab
Clarify bitfile version print
