### 2022-08-19 8879c36
Fix FIFO index and sample size in DDR3 parameters
