### 2022-08-23 1b7d68f
Add append option to DDR3.save_data
