### 2022-08-23 3378e0e
Return new data instead of all data
