### 2022-08-29 097b3d2
Bitfile versioning and DDR3.save_data append (#18)
* Add bitfile_version to FPGA class

* Add bitfile_version to .h5 header in DDR3.save_data

* Add str_bitfile_version

* data_to_names accounts for bitfile version

* Fix data_to_names bitfile_version default value bug

* Account for KeyError when accessing BITFILE_VERSION Endpoint

* Clarify bitfile version print

* Fix FIFO index and sample size in DDR3 parameters

* Add append option to DDR3.save_data

* Fix append option

* Fix append option

* Fix append option

* Add full data return to DDR3.save_data

* Return new data instead of all da
[…]
