+++ python/src/pyripherals/peripherals/DDR3.py
@@ -51,25 +51,23 @@ class DDR3():
     
     """
 
-    def __init__(self, fpga, endpoints=None, data_version='ADC_NO_TIMESTAMPS'):
+    def __init__(self, fpga, endpoints=None, data_version='TIMESTAMPS'):
         if endpoints is None:
             endpoints = Endpoint.get_chip_endpoints('DDR3')
         self.fpga = fpga
         self.endpoints = endpoints
         self.parameters = {'BLOCK_SIZE': 2048,  # 1/2 the incoming FIFO depth in bytes (size of the BlockPipeIn)
-                           'sample_size': 65536,  # per channel
                            # number of channels that the DDR is striped between (for DACs)
                            'channels': 8,
                            'update_period': 400e-9,  # 2.5 MHz -- requires SCLK ~ 50 MHZ
-                           'port1_index': 0x7f_ff_f8,
+                           'port1_index': 0x3_7f_ff_f8,
                            'adc_channels': 8,  # number of 2 byte chunks in DDR
                            'adc_period': 200e-9
                            }
 
         # the index is the DDR address that the circular buffer stops at.
         # need to write all the way up to this stoping point otherwise the SPI output will glitch
-        self.parameters['sample_size'] = int(
-            (self.parameters['port1_index'] + 8)/2)
+        self.parameters['sample_size'] = int((self.parameters['port1_index'] + 8)/4)
         self.parameters['data_version'] = data_version  # sets deswizzling mode
 
         self.data_arrays = []
@@ -602,7 +600,7 @@ class DDR3():
 
         return chan_data
 
-    def data_to_names(self, chan_data, old=False):
+    def data_to_names(self, chan_data, bitfile_version=None):
         """
         Put deswizzled data into dictionaries with names that match with the data sources. 
         Complete twos complement conversion where necessary. Check timestamps for skips.
@@ -617,8 +615,8 @@ class DDR3():
         ----------
         chan_data : dict of np.arrays
             data from reading DDR (minimally processed into 2 byte containers)
-        old : bool
-            indicates if the chan data is from before the ADS channel B was repositioned to fix synchronization issues
+        bitfile_version : int
+            The bitfile version that the .h5 file which held chan_data was created with. Defaults to the bitfile version of the FPGA
         
         Returns
         -------
@@ -641,6 +639,14 @@ class DDR3():
                or the timestamp steps are not all the same
         """
 
+        if bitfile_version is None:
+            # Old code, hasn't been updated to pass bitfile_version from .h5 file header
+            # Default to FPGA bitfile version, or version 1 if the FPGA bitfile version is None
+            if self.fpga.bitfile_version is None:
+                bitfile_version = 1
+            else:
+                bitfile_version = self.fpga.bitfile_version
+
         # first version of ADC data before DACs + timestamps are stored
         if self.parameters['data_version'] == 'ADC_NO_TIMESTAMPS':
             adc_data = chan_data
@@ -657,7 +663,7 @@ class DDR3():
                 # adc_data[i] = custom_signed_to_int(chan_data[i], 16)
                 adc_data[i] = chan_data[i]
 
-            if old:
+            if bitfile_version < 2: # 00.00.02 -> 2
                 lsb = chan_data[6][0::5].astype(np.uint64)
             else:
                 lsb = chan_data[7][1::5].astype(np.uint64)
@@ -686,7 +692,7 @@ class DDR3():
 
             ads = {}
             ads['A'] = custom_signed_to_int(chan_data[7][0::5], 16)
-            if old:
+            if bitfile_version < 2: # 00.00.02 -> 2
                 ads['B'] = custom_signed_to_int(chan_data[7][1::5], 16)
             else:
[…]
