### 2022-09-02 b9713a0
Fix bug where end of last save gets cut off and not all data is returned (#20)
