### 2022-09-11 #4586 @danielskatz
👋 @sptennak - Would you be willing to review this JOSS submission?
