### 2022-09-13 307de12
Check types in custom_signed_to_int and int_to_custom_signed and raise TypeErrors when not int
