### 2022-09-14 #4762 @danielskatz
👋 @Ajstros (Abraham Stroschein) - please work on the possibly missing DOI and incorrectly prefixed DOI that editorialbot suggests, but note that the missing one may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
