+++ paper/paper.bib
@@ -84,7 +84,7 @@
 	title = {{PyMeasure}},
 	url = {https://pymeasure.readthedocs.io/en/latest/},
 	author = {Jermain, Colin and Rowlands, Graham},
-	doi = {https://doi.org/10.5281/zenodo.3732545},
+	doi = {10.5281/zenodo.3732545},
 	year = {2020}
 }
 
