### 2022-10-01 1a4ebdf
Change get_chip_endpoints paragraph to fit
