### 2022-10-01 1951d0d
Fix paper line length (#21)
* Keep lines under 80 characters

* Change get_chip_endpoints paragraph to fit
