### 2022-10-02 Ajstros/pyripherals#22 @askuric
[JOSS REVIEW] tests no_fpga failing
**Describe the bug**
I'm testing the installation of the pyripherals pip package. I've followed the instalation guidelines, and installed it from pip. However since I do not have FPGA to test it with and I'm working on linux, I cannot install the FronPannel API.  So the first test of
```python
from pyripherals.utils import create_yaml
create_yaml()
```
was successful.

However when I try to run the tests, the line `python -m pytest -m no_fpga`gives this error:
```
================================================================================== test session starts ==================================================================================
rootdir: ../joss/pyripherals, configfile: pyproject.toml
collected 1339 items / 1 error / 515 deselected / 824 selected

======================================================================================== ERRORS =========================================================================================
_________________________________________________________________ ERROR collecting python/tests/unit_tests/i2c_test.py __________________________________________________________________
python/tests/unit_tests/i2c_test.py:20: in <module>
    message_eps = Endpoint.get_chip_endpoints('I2CTEST')
../../python3.10/site-packages/pyripherals/core.py:335: in get_chip_endpoints
    Endpoint.update_endpoints_from_defines()
../../python3.10/site-packages/pyripherals/core.py:159: in update_endpoints_from_defines
    with open(ep_defines_path, 'r') as file:
E   TypeError: expected str, bytes or os.PathLike object, not NoneType
================================================================================ short test summary info ================================================================================
ERROR python/tests/unit_tests/i2c_test.py - TypeError: expected str, bytes or os.PathLike object, not NoneType
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 1 error during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
=========================================================================== 515 deselected, 1 error in 1.51s ============================================================================
```
I have cloned your repository and I have the same behavior for the master and v0.0.2 branch.

I am not sure if this is due to not having FronPannel API, or maybe Linux OS or something else.
