### 2022-10-02 #4762 @askuric
Hi everyone,
My review of Pyripherals  is here:

## Pyripherals

Pyripherals is  a python pip package facilitating the communication to different FPGA based peripheral devices. It provides an user-friendly abstraction layer to different communication protocols and implements interfaces to many different standard FPGA based data acquisition devices. This all makes Pyripherals an useful tool for setting up an experiments requiring real-time data exchange.

## Paper comments
Text:
- line 12:
"yet the Python developments are generally useful to interface to electronic chips containing registers."
    - Maybe a bit too general, as all the electronics chips and microcontrollers contain registers at some level. Maybe you could you be a bit more specific and say that it is for the "communication".
- line 37:
" The Opal Kelly XEM7310 FPGA that we use for communication controllers to demonstrate pyripherals is common in research environments which allows other labs to accelerate their development of FPGA to electronic chip interfaces using pyripherals."
    - Some commas are probably necessary here :D
- table page 2:
  - It would be maybe good to have a caption and to explain each one of the table entries to the reader, as not everyone will be able to understand those. And maybe refer how the name 'ABC012' corresponds to the table entries, if it does. If it doesn't, maybe it would be a good idea to use a more descriptive name that 'ABC012', like 'MyPeripheral' or something like that.
- line 69:
"The line below shows an example that defines an endpoint named “WRITE_IN” that belongs to peripheral “ABC012” with an address of 0x04 and a bit_width of 32 that adds 7 to the address every time it is advanced."
  - I would suggest to add a sentence or two before this line (or after the line 65) to explain what the endpoint actually is and how it is defined. It would probably be easier to understand each of the parameters than learning it directly from the example.
  - I'd also suggest to change the name of the endpoint to something more descriptive, "WRITE_IN" an input I imagine, but an input of the FPGA or the PC?
  - Also, maybe it would be a good idea to mention that the 8'h  means 8 bit hex number for the sake of clarity, even though it is described in the documentation.
- line 86:
"Using the Endpoint class in pyripherals with a definitions file extends the capabilities of the Opal Kelly FrontPanel API by automatically linking the Python and Verilog endpoint data.
  - Few commas missing, maybe also repharse it to be more clear.
- Something to consider
    - Maybe it would be a good idea to add a paragraph at the begging of the Summary section explaining in broad terms the functioanllity of the code and the general stracture. We have a PC, a FPGA and the communication protoclo which has registers and endpoints.
    - Also there you coul add a schematic of the peripheral communication showing a PC, FPGA, the communication and maybe endpoints or registers, I'm not sure how it should look exactly. But a visual could help in understanding better all the components.

## Documentation commnets

- tests:
    - the line `py -m pytest -m usable`, I am not sure if it's just me by `py` does not work only with `pytest` installed. I'd suggest to use `python -m pytest -m usable`.
    - I've also have one of the tests failing: https://github.com/Ajstros/pyripherals/issues/22
    - You could also look into setting up the testing procedure using github CI (the github actions). That way every github commit would be automatically tested.
- Community guidelines:
  - You have a very nice guide for contributing the new peripheral.
  - However it might be a good idea to make it more general, if people are motivated to contribute by adding a functionality to the existing code, correcting bugs, improve the code efficiency or any other form of contribution.

## Code comments

- The code is well structured and written, and the unit tests are provided.
-  I was not able to test the functionality of the code as I do not have the hardware necessary.
- The installation procedure is well documented and I had no trouble installing the package and running simple examples.
