### 2022-10-06 Ajstros/pyripherals#22 @lucask07
Hi @askuric
Could you post your YAML configuration file (described [here](https://pyripherals.readthedocs.io/en/latest/installation.html#installation-yaml))? The error output suggests that _ep_defines_path_ is None.

You are correct that we have not been able to test with Linux OS. But we have verified with both Windows and MAC OS.

My MAC configuration file is at ~/.pyripherals/config.yaml

> endpoint_max_width: 32
>   ep_defines_path: /Users/koer2434/Documents/covg_fpga/fpga_XEM7310/fpga_XEM7310.srcs/sources_1/ep_defines.v
>   fpga_bitfile_path: /Users/koer2434/Documents/covg_fpga/fpga_XEM7310/fpga_XEM7310.runs/impl_1/top_level_module.bit
>   frontpanel_path: /Users/koer2434/Documents/frontpanel_5p26
>   registers_path: /Users/koer2434/Documents/covg_fpga/python/Registers.xlsx
