### 2022-10-06 Ajstros/pyripherals#22 @untzag
I guess we are in a somewhat strange place, we would like to know the minimal path to run tests without hardware
