### 2022-10-06 #4762 @untzag
Pyripherals seems like a great project and it's clearly a powerful tool which has already been used to drive a variety of experiments. I'm thinking I need to buy an OpalKelly FPGA! I recommend publication and congrats on the great work.

I think pyripherals needs one more round of documentation work with an eye towards high-level explanation for outsiders wanting to start from scratch. You've built a system with a lot of moving parts, and it's opaque to me as a reader how those fit together into an experimental control system. I think you can expect a basic understanding of busses, registers, and the concept of a Python API---but readers like me need help with the layers between.
