+++ paper/paper.md
@@ -25,7 +25,7 @@ bibliography: paper.bib
 ---
 
 # Statement of Need
-We are developing a data acquisition system (DAQ) for real-time feedback that uses FPGA-based control of and acquisition from various electronic chips, or peripherals. Because these peripherals communicate over multiple protocols (SPI, I2C, LVDS) through an FPGA, we designed *pyripherals* to organize and abstract registers, the communication protocol, and the host computer interface to each communication controller. The software and firmware are designed for Opal Kelly FPGA modules, yet the Python developments are generally useful to interface to electronic chips containing registers. 
+We are developing a data acquisition system (DAQ) for real-time feedback that uses FPGA-based control of and acquisition from various electronic chips, or peripherals. Because these peripherals communicate over multiple protocols (SPI, I2C, LVDS) through an FPGA, we designed *pyripherals* to organize and abstract registers, the communication protocol, and the host computer interface to each communication controller. The software and firmware are designed for Opal Kelly FPGA modules, yet the Python developments are generally useful for organizing the data fields of electronic chips, especially for communication.
 
 Data from a register is accessed with the address of the register and the bit indices of the data field, but users will often refer to the data field by its name, which *pyripherals* retains as a dictionary key. When passed the location of the data field, parameterized functions automatically format the data or command as required by the communication interface used by the peripheral. The assembled message is passed to the appropriate hardware controller responsible for low-level communication with the peripheral. In this solution, the addressing, bit indexing, and formatting are handled by *pyripherals* before the message is sent over the Opal Kelly FrontPanel API to a hardware-level communication controller on an Opal Kelly FPGA [@opalkelly.comFrontPanel], which handles low-level communication with the desired peripheral.
 
@@ -35,7 +35,7 @@ This DAQ system and the *pyripherals* software are crucial components of a digit
 
 ## Registers
 
-*pyripherals* reads an Excel spreadsheet that holds the name, address, lower bit index, upper bit index, and bit width of each data field. Different peripherals can be separated into different sheets within the register index. The table below shows an example sheet for a peripheral named “ABC012”.
+*pyripherals* reads an Excel spreadsheet that holds the name, address, lower bit index, upper bit index, and bit width of each data field. Different peripherals can be separated into different sheets within the register index. The table below shows an example sheet for a peripheral named “TestPeripheral”.
 
 | Name          | Hex Address | Default Value | Bit Width | Bit Index (High) | Bit Index (Low) |
 | ------------- | ----------- | ------------- | --------- | ---------------- | --------------- |
@@ -44,18 +44,25 @@ This DAQ system and the *pyripherals* software are crucial components of a digit
 | CONFIG        | 0x01        | 0xff          | 8         | 7                | 0               |
 | ID            | 0x01        | 0x0           | 8         | 15               | 8               |
 
+- Name = name of the data field chosen by the user. This is how the register will be accessed in Python.
+- Hex Address = address of the register on the chip as given by its datasheet. Must be formatted as a hexadecimal number with "0x" prefix.
+- Default Value = the default value of the register as given by the chip's datasheet.
+- Bit Width = the number of bits in the data field.
[…]
