### 2022-10-11 #4762 @lucask07
@untzag, @danielskatz, @askuric
Thank you so much for the great feedback and we really appreciate your time! We (@Ajstros) are working through these and will get back to you soon. Sorry for the slow response. One of us is an undergraduate student taking too many courses and the other is trying to teach 3 courses.

I completely agree that there are many moving parts. We will take a careful look at the paper and the documentation to try to help new users get started.

Pyripherals is a general Python solution to organize and interface to chips with registers but in our lab, we exclusively use it to interface with the FPGA design here:
[https://github.com/lucask07/covg_fpga](https://github.com/lucask07/covg_fpga)
