### 2022-10-11 #4762 @lucask07
>

> Quoting from @askuric
>
> > Maybe it would be a good idea to add a paragraph at the begging of the Summary section explaining in broad terms the functioanllity of the code and the general stracture. We have a PC, a FPGA and the communication protoclo which has registers and endpoints.
>
> As somebody who is a little bit further away from the FPGA world, I strongly advise you to add some more description here. Your statement of need is motivating---I would LOVE to have microsecond latency real-time feedback in some of the instruments I'm responsible for. Personally I cannot discern the big picture of how pyripherals fits into this system.
>
> Here's my best attempt to describe what we have here. I'm wrong I'm sure, but perhaps my lack of understanding will be illustrative 😄
>
>     * users create register spreadsheets corresponding to IC internals
Correct
>
>     * all ICs are connected to FPGA
>
Yes. This may require a custom circuit board that the ICs are mounted to (e.g. [https://github.com/lucask07/open_covg_daq_pcb](https://github.com/lucask07/open_covg_daq_pcb)). An ADC evaluation board or for example a [ToF sensor eval board](https://www.digikey.com/en/products/detail/mikroelektronika/MIKROE-4592/14306689?s=N4IgTCBcDaILYEsDWAnA9gUwLQBYCsAnBALoC%2BQA).

>     * pyripherals somehow tells the FPGA about which ICs are connected to which buses, and which registers those ICs have
There is a file (ep_defines.v) that is shared between Verilog and pyripherals. This file specifies the addresses of each communication controller (e.g. SPI controller) that is instantiated in the FPGA. The FPGA does not know which registers the IC has. Pyripherals knows the registers and sends the message through the FPGA to the IC.
>
>     * does this involve verilog code generation?
Yes, this requires verilog but not verilog "code generation".
>
>     * pyripherals offers a python interface to the ICs where the FPGA somehow acts as a host bus adapter
Yes, and the FPGA is a host bus adapter for multiple interfaces (SPI, I2C, LVDS, etc.).
>
>
> Presumably the host python <-> FPGA bus is not particularly real-time, so where does my real-time feedback live?

In the FPGA via digital signal processing. ADC data can be scaled and filtered and then drive DAC(s). You are correct that this does not go through the python <-> FPGA bus. Real-time feedback is a motivator for an FPGA and peripherals (ADCs and DACs) but not a motivator directly for pyripherals. For real-time feedback experiments pyripherals first initializes ADCs (sets pre-amp gain or filter bandwidth) and DACs before control is handed off to the FPGA.

>
> Honestly as I write this out it gets more confusing. Again I'm not in the FPGA world so I'm going to need it spelled out slowly. I think new users of pyripherals might be in a similar place.
Totally agree, we have some documentation work to do!
