### 2022-10-15 65c4b4d
Fix calc_impedance_test by increasing test frequency
