### 2022-10-15 e56e149
Change all DDR3 parameters to be class attributes
Allows DDR3 parameters to be accessed without instantiating DDR3
