+++ python/src/pyripherals/peripherals/DDR3.py
@@ -51,27 +51,27 @@ class DDR3():
     
     """
 
+    BLOCK_SIZE = 2048,  # 1/2 the incoming FIFO depth in bytes (size of the BlockPipeIn)
+    # number of channels that the DDR is striped between (for DACs)
+    NUM_CHANNELS = 8,
+    UPDATE_PERIOD = 400e-9,  # 2.5 MHz -- requires SCLK ~ 50 MHZ
+    PORT1_INDEX = 0x3_7f_ff_f8,
+    NUM_ADC_CHANNELS = 8,  # number of 2 byte chunks in DDR
+    ADC_PERIOD = 200e-9
+
+    # the index is the DDR address that the circular buffer stops at.
+    # need to write all the way up to this stoping point otherwise the SPI output will glitch
+    SAMPLE_SIZE = int((PORT1_INDEX + 8)/4)
+
     def __init__(self, fpga, endpoints=None, data_version='TIMESTAMPS'):
         if endpoints is None:
             endpoints = Endpoint.get_chip_endpoints('DDR3')
         self.fpga = fpga
         self.endpoints = endpoints
-        self.parameters = {'BLOCK_SIZE': 2048,  # 1/2 the incoming FIFO depth in bytes (size of the BlockPipeIn)
-                           # number of channels that the DDR is striped between (for DACs)
-                           'channels': 8,
-                           'update_period': 400e-9,  # 2.5 MHz -- requires SCLK ~ 50 MHZ
-                           'port1_index': 0x3_7f_ff_f8,
-                           'adc_channels': 8,  # number of 2 byte chunks in DDR
-                           'adc_period': 200e-9
-                           }
-
-        # the index is the DDR address that the circular buffer stops at.
-        # need to write all the way up to this stoping point otherwise the SPI output will glitch
-        self.parameters['sample_size'] = int((self.parameters['port1_index'] + 8)/4)
-        self.parameters['data_version'] = data_version  # sets deswizzling mode
+        self.data_version = data_version  # sets deswizzling mode
 
         self.data_arrays = []
-        for i in range(self.parameters['channels']):
+        for i in range(DDR3.NUM_CHANNELS):
             self.data_arrays.append(np.zeros(
                 self.parameters['sample_size']).astype(np.uint16))
 
@@ -133,7 +133,7 @@ class DDR3():
             The closest possible frequency
         """
 
-        samples_per_period = (1/freq) / self.parameters['update_period']
+        samples_per_period = (1/freq) / DDR3.UPDATE_PERIOD
 
         if samples_per_period <= 2:
             print('Frequency is too high for the DDR update rate')
@@ -144,7 +144,7 @@ class DDR3():
         round_samples_per_period = self.parameters['sample_size'] / \
             round_total_periods
         new_frequency = 1 / \
-            (self.parameters['update_period'] * round_samples_per_period)
+            (DDR3.UPDATE_PERIOD * round_samples_per_period)
 
         return new_frequency
 
@@ -181,8 +181,8 @@ class DDR3():
         if actual_frequency:
             frequency = self.closest_frequency(frequency)
 
-        t = np.arange(0, self.parameters['update_period']*self.parameters['sample_size'],
-                      self.parameters['update_period'])
+        t = np.arange(0, DDR3.UPDATE_PERIOD*self.parameters['sample_size'],
+                      DDR3.UPDATE_PERIOD)
         # print('length of time axis after creation ', len(t))
         ddr_seq = (amplitude)*np.sin(t*frequency*2*np.pi) + offset
         if any(ddr_seq < 0) or any(ddr_seq > (2**16-1)):
@@ -269,10 +269,10 @@ class DDR3():
         """Write the channels as striped data to the DDR."""
 
[…]
