### 2022-10-15 cfa488a
Turn "make_..." style methods into staticmethods
Allows calling method from the class or an instance
