### 2022-10-15 108a95b
Make DDR3 class attributes separate instead of tuple
