### 2022-10-15 bbdeaf4
Update calc_impedance_test to use DDR3 without instantiation
