### 2022-10-20 2802c1d
Add Python app workflow to lint and test
