+++ .github/workflows/python-app.yml
@@ -27,7 +27,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install flake8 pytest
-        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+        pip install git+https://github.com/Ajstros/pyripherals@${{ github.reg_name }}
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
