### 2022-10-20 65e66c4
Add environment variable to hold github.ref_name context
