+++ .github/workflows/python-app.yml
@@ -11,6 +11,9 @@ on:
 
 permissions:
   contents: read
+  
+env:
+  BRANCH: ${{ github.reg_name }}
 
 jobs:
   build:
@@ -27,7 +30,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install flake8 pytest
-        pip install git+https://github.com/Ajstros/pyripherals@${{ github.reg_name }}
+        pip install git+https://github.com/Ajstros/pyripherals@$BRANCH
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
