### 2022-10-20 1ab5c24
Make pip install location a string
I am hoping this will allow the $BRANCH environment variable to be used
