+++ .github/workflows/python-app.yml
@@ -30,7 +30,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install flake8 pytest
-        pip install git+https://github.com/Ajstros/pyripherals@$BRANCH
+        pip install "git+https://github.com/Ajstros/pyripherals@$BRANCH"
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
