### 2022-10-20 3d519c5
Fix ref_name typo and add echo
Echo to display the branch we are installing from (should be the branch the commit is on)
