+++ .github/workflows/python-app.yml
@@ -13,7 +13,7 @@ permissions:
   contents: read
   
 env:
-  BRANCH: ${{ github.reg_name }}
+  BRANCH: ${{ github.ref_name }}
 
 jobs:
   build:
@@ -30,6 +30,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install flake8 pytest
+        echo "Installing from branch: $BRANCH"
         pip install "git+https://github.com/Ajstros/pyripherals@$BRANCH"
     - name: Lint with flake8
       run: |
