### 2022-10-20 f5b7686
Use github.head_ref instead of github.ref_name
