### 2022-10-20 eb5afcf
Fix Python App Workflow and no_fpga tests (#25) Fixes #22
* Fix calc_impedance_test by increasing test frequency

* Allow for no ep_defines path, exception on file not existing

* Change all DDR3 parameters to be class attributes

Allows DDR3 parameters to be accessed without instantiating DDR3

* Turn "make_..." style methods into staticmethods

Allows calling method from the class or an instance

* Make DDR3 class attributes separate instead of tuple

* Update calc_impedance_test to use DDR3 without instantiation

* Attempt to pip install from GitHub branch

* Add environment variable to hol
[…]
