+++ .github/workflows/python-app.yml
@@ -11,6 +11,9 @@ on:
 
 permissions:
   contents: read
+  
+env:
+  BRANCH: ${{ github.head_ref }}
 
 jobs:
   build:
@@ -27,7 +30,8 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install flake8 pytest
-        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+        echo "Installing from branch: $BRANCH"
+        pip install "git+https://github.com/Ajstros/pyripherals@$BRANCH"
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
