### 2022-10-20 e3f005d
Use main for branch name on commits to main
According to example from GitHub: https://docs.github.com/en/actions/examples/using-concurrency-expressions-and-a-test-matrix
