### 2022-10-20 #4762 @Ajstros
@untzag @askuric Sorry for the delay, but the [testing issue](https://github.com/Ajstros/pyripherals/issues/22) with `no_fpga` tests has been fixed. You should now be able to run those tests with or without a `config.yaml` file regardless of the `ep_defines_path` value. These tests have also been added to a GitHub workflow run on any commit or pull request to the main branch.
