### 2022-10-27 65aea96
Add abstract paragraph
