### 2022-10-27 71c71f1
add figure for JOSS paper
