### 2022-10-27 c9947fc
Add get_serial_number link
