### 2022-10-27 377be5b
Reword figure 1 caption
