### 2022-10-27 af333d0
Remove abstract paragraph
