### 2022-10-27 #4762 @lucask07
Hi @danielskatz
We have completed our code fixes and paper/documentation upgrades in response to the feedback from the reviewers. The brief summary is:

1. The "no_fpga" tests are working and should allow the reviewers to run a subset of tests without an FPGA. (GitHub Issue: [JOSS Review: tests no_fpga failing](https://github.com/Ajstros/pyripherals/issues/22) )
2. Continuous integration (CI) was added that runs tests on a push to GitHub.
3. A figure was created for the paper to explain the system and need at a high-level. This leads off the 'Summary' section after which the paper goes into the details. (this is intended to address the "a lot of moving parts" comment of @untzag and the request for a better Summary from @askuric).
4. Specific paper comments were addressed.

> > Similar DAQ systems (Leibrandt & Heidecker, 2015; Yu et al., 2018) create MHz bandwidth servos for physics experiments but these works do not expose the host software.
>
> To me, it's unclear what you mean by "expose" here. I think you mean that their host software is not open source or freely available? I think for me it's down to the word "expose", consider choosing a different word.

This has been rewritten to "but the host software is either not publicly available or not generalizable."
