### 2022-10-27 #4762 @lucask07
> Quoting again from @askuric
>
> > You could also look into setting up the testing procedure using github CI (the github actions). That way every github commit would be automatically tested.
>
> In my opinion this should be blocking for publication. This is a good opportunity to get real CI set up and running.
>
> I understand that many of these code paths aren't possible to run without hardware, but you do have `no_fpga` tests. Even just showing that the package can be installed and imported with CI is useful for dealing with potential chaos from new contributors.
>
> As long as you're working on CI, consider adding a workflow for publishing to PyPI---that's a great feature for a package that's going to get worked on by multiple people.

Great idea, CI for tests upon on push to GitHub have been setup.
