### 2022-10-27 #4762 @lucask07
> > Similar DAQ systems (Leibrandt & Heidecker, 2015; Yu et al., 2018) create MHz bandwidth servos for physics experiments but these works do not expose the host software.
>
> To me, it's unclear what you mean by "expose" here. I think you mean that their host software is not open source or freely available? I think for me it's down to the word "expose", consider choosing a different word.

Good point. This has been rewritten to "but the host software is either not publicly available or not generalizable."
