+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = pyripherals
-version = 0.0.2
+version = 0.0.3
 author = Abe Stroschein
 author_email = ajstros@duck.com
 description = Python solution for communicating with peripheral ICs
