### 2022-11-09 #4762 @Ajstros
@danielskatz the new release is v0.0.3. I have uploaded to Zenodo and checked the metadata. The DOI is 10.5281/zenodo.7308636
